% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phen.R
\name{sim.phen}
\alias{sim.phen}
\title{Simulate genetic data.}
\usage{
sim.phen(
  N = 1e+05,
  alpha = 0,
  beta = 1,
  rho = 0.5,
  rho_g = 0.7,
  gamma = 1,
  mu_dir = 0,
  sd_dir = 1,
  mu_nur = 0,
  sd_nur = NULL,
  sd_ind = 1,
  sd_fam = 1,
  het = "none"
)
}
\arguments{
\item{N}{Number of simulated sibling pairs}

\item{rho}{Genetic correlation between siblings.}

\item{mu_dir}{Mean direct genetic effect.}

\item{sd_dir}{Standard deviation of direct genetic effect.}

\item{mu_nur}{Mean genetic nuture effect.}

\item{sd_nur}{Standard deviation of genetic nurture effect.}

\item{het}{Specify type of heteroskedasticity. Can be none, individual, family, or full.}
}
\value{
A dataframe of sibling pairs, genotypes, and phenotypes.
}
\description{
Simulate genetic data.
}
\examples{
sim.phen(N=1e5, alpha=0, beta=1, rho=0.5, rho_g=0.7, gamma=1, mu_dir=0, sd_dir=1, mu_nur=0, sd_nur=NULL, sd_ind=1, sd_fam=1, het = 'none'
}
